+++ drivers/net/wireless/realtek/rtlwifi/rtl8192ee/hw.h
@@ -37,5 +37,5 @@ void rtl92ee_suspend(struct ieee80211_hw *hw);
 void rtl92ee_resume(struct ieee80211_hw *hw);
 void rtl92ee_allow_all_destaddr(struct ieee80211_hw *hw, bool allow_all_da,
 				bool write_into_reg);
-void rtl92ee_fw_clk_off_timer_callback(unsigned long data);
+void rtl92ee_fw_clk_off_timer_callback(uintptr_t data);
 #endif
